built fileFinder function, built parcelLister function
fileFinder function allows the program to detect files within the same directory and automatically identify drone and parcel files! this way no examining teacher has to tamper with the code for it to work, and the program can simply be run in the same directory as the parcel and drone files to work.